Add `InspectableRun.inputs` API.
